refactor: expand team management domain to users
