+++ src/domains/team-management/team-management.routes.tsx
@@ -1,8 +1,15 @@
 import { lazy } from "react";
 import type { RouteObject } from "react-router-dom";
+import { Navigate } from "react-router-dom";
 import AuthGuard from "@/guards/AuthGuard";
+import TeamMakerGuard from "@/guards/TeamMakerGuard";
 
-const TeamManagement = lazy(() => import("./pages/common/TeamManagement"));
+const TeamManagementUsers = lazy(
+	() => import("./pages/users/TeamManagementUsers"),
+);
+const TeamManagementCategory = lazy(
+	() => import("./pages/category/TeamManagementCategory"),
+);
 
 //
 //
@@ -10,9 +17,13 @@ const TeamManagement = lazy(() => import("./pages/common/TeamManagement"));
 
 /**
  * @property {string} ROOT `/team-management`
+ * @property {string} USERS `/team-management/users`
+ * @property {string} CATEGORY `/team-management/category`
  */
 export const TEAM_MANAGEMENT_ROUTE_PATH = {
 	ROOT: "/team-management",
+	USERS: "/team-management/users",
+	CATEGORY: "/team-management/category",
 };
 
 //
@@ -22,10 +33,22 @@ export const TEAM_MANAGEMENT_ROUTE_PATH = {
 export const teamManagementRouter: RouteObject[] = [
 	{
 		path: TEAM_MANAGEMENT_ROUTE_PATH.ROOT,
+		element: <Navigate to={TEAM_MANAGEMENT_ROUTE_PATH.USERS} replace />,
+	},
+	{
+		path: TEAM_MANAGEMENT_ROUTE_PATH.USERS,
 		element: (
 			<AuthGuard>
-				<TeamManagement />
+				<TeamManagementUsers />
 			</AuthGuard>
+		),
+	},
+	{
+		path: TEAM_MANAGEMENT_ROUTE_PATH.CATEGORY,
+		element: (
+			<TeamMakerGuard>
+				<TeamManagementCategory />
+			</TeamMakerGuard>
 		),
 	},
 ];
